requirements/dev.txt
# for bin/generate-vendor-lib.py
virtualenv==1.7.1.2

# for testing
py==1.4.9
pytest==2.2.4
pytest-django==1.3.2
pytest-localserver==0.1.4
factory_boy==1.1.4
